Remove `useLocale` and friend from API (#711)
