+++ packages/react-components/src/components/utils/testUtils.tsx
@@ -4,7 +4,7 @@
 import React from 'react';
 import { mount, ReactWrapper, shallow, ShallowWrapper } from 'enzyme';
 import { LocalizationProvider, Locale, ComponentStrings } from '../../localization/LocalizationProvider';
-import en_US from '../../localization/translated/en_US.json';
+import { en_US } from '../../locales';
 import { PartialDeep } from 'type-fest';
 
 export const mountWithLocalization = (node: React.ReactElement, locale: Locale): ReactWrapper => {
@@ -22,7 +22,7 @@ export const shallowWithLocalization = (node: React.ReactElement, locale: Locale
 };
 
 export const createTestLocale = (testStrings: PartialDeep<ComponentStrings>): Locale => {
-  const strings: ComponentStrings = en_US;
+  const strings: ComponentStrings = en_US.strings;
   Object.keys(testStrings).forEach((key: string) => {
     strings[key] = { ...strings[key], ...testStrings[key] };
   });
