+++ packages/react-components/src/index.ts
@@ -2,9 +2,11 @@
 // Licensed under the MIT license.
 
 export * from './components';
-export * from './localization';
-export * from './theming';
 export * from './identifiers';
+export * from './locales';
+export { LocalizationProvider } from './localization';
+export type { ComponentStrings, Locale, LocalizationProviderProps } from './localization';
+export * from './theming';
 
 export type {
   BaseCustomStylesProps,
