+++ packages/react-components/src/locales/translated/de_DE.json
