+++ packages/react-components/src/locales/translated/en_US.json
