+++ packages/react-components/src/locales/translated/fr_FR.json
